@@ -9,8 +9,8 @@ import com.fasterxml.jackson.databind.JsonNode;
 import io.mosip.esignet.mock.identitysystem.util.ErrorConstants;
 import lombok.Data;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
 import java.util.Map;
 
 
