@@ -8,11 +8,12 @@ package io.mosip.esignet.mock.identitysystem.entity;
 import io.mosip.esignet.mock.identitysystem.dto.Valid;
 import lombok.*;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import java.time.LocalDateTime;
+
 
 @AllArgsConstructor
 @NoArgsConstructor
