@@ -3,8 +3,8 @@ package io.mosip.esignet.mock.identitysystem.validator;
 
 import io.mosip.esignet.mock.identitysystem.util.ErrorConstants;
 
-import javax.validation.Constraint;
-import javax.validation.Payload;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
