@@ -1,21 +1,20 @@
 package io.mosip.esignet.mock.identitysystem.util;
 
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.cache.Cache;
 import org.springframework.cache.CacheManager;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class CacheUtilServiceTest {
     @InjectMocks
     private CacheUtilService cacheUtilService;
@@ -27,10 +26,11 @@ public class CacheUtilServiceTest {
     private CacheManager cacheManager;
 
 
-    @Before
+    @BeforeEach
     public void setUp(){
         when(cacheManager.getCache(Constants.TRANSACTION_HASH)).thenReturn(cache);
     }
+
     @Test
     public void test_setTransactionHash() {
         String transactionHash = "testTransactionHash";
@@ -43,7 +43,7 @@ public class CacheUtilServiceTest {
         String transactionHash = "testTransactionHash";
         when(cache.get(transactionHash, String.class)).thenReturn(transactionHash);
         boolean result = cacheUtilService.getTransactionHash(transactionHash);
-        assertTrue(result);
+        Assertions.assertTrue(result);
     }
 
     @Test
@@ -51,7 +51,7 @@ public class CacheUtilServiceTest {
         String transactionHash = "testTransactionHash";
         when(cache.get(transactionHash, String.class)).thenReturn(null);
         boolean result = cacheUtilService.getTransactionHash(transactionHash);
-        assertFalse(result);
+        Assertions.assertFalse(result);
     }
 
     @Test
