@@ -11,18 +11,17 @@ import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.mosip.esignet.mock.identitysystem.dto.LanguageValue;
 import io.mosip.esignet.mock.identitysystem.exception.MockIdentityException;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.List;
 
-import static org.junit.Assert.*;
-
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class HelperUtilTest {
 
     private static final String TEST_EMAIL = "test@example.com";
@@ -32,28 +31,28 @@ public class HelperUtilTest {
     @Test
     public void b64EncodeString_validDetails_thenPass() {
         String encoded = HelperUtil.b64Encode(TEST_STRING);
-        assertNotNull(encoded);
-        assertEquals("SGVsbG8gV29ybGQ", encoded);
+        Assertions.assertNotNull(encoded);
+        Assertions.assertEquals("SGVsbG8gV29ybGQ", encoded);
     }
 
     @Test
     public void b64EncodeByteArray_validDetails_thenPass() {
         byte[] bytes = TEST_STRING.getBytes(StandardCharsets.UTF_8);
         String encoded = HelperUtil.b64Encode(bytes);
-        assertNotNull(encoded);
-        assertEquals("SGVsbG8gV29ybGQ", encoded);
+        Assertions.assertNotNull(encoded);
+        Assertions.assertEquals("SGVsbG8gV29ybGQ", encoded);
     }
 
     @Test
     public void maskMobile_validMobile_thenPass() {
         String masked = HelperUtil.maskMobile(TEST_MOBILE);
-        assertEquals("XXXXXX7890", masked);
+        Assertions.assertEquals("XXXXXX7890", masked);
     }
 
     @Test
     public void maskEmail_validEmail_thenPass() {
         String masked = HelperUtil.maskEmail(TEST_EMAIL);
-        assertEquals("XXsX@example.com", masked);
+        Assertions.assertEquals("XXsX@example.com", masked);
     }
 
     @Test
@@ -66,7 +65,7 @@ public class HelperUtilTest {
         ((ObjectNode) langNode).set("language", JsonNodeFactory.instance.textNode("en"));
         ((ObjectNode) langNode).set("value", JsonNodeFactory.instance.textNode("Value in English"));
         String result = HelperUtil.getIdentityDataValue(jsonNode, "testField", "en");
-        assertEquals("Value in English", result);
+        Assertions.assertEquals("Value in English", result);
     }
 
     @Test
@@ -77,27 +76,27 @@ public class HelperUtilTest {
         langNode.set("language", JsonNodeFactory.instance.textNode("en"));
         langNode.set("value", JsonNodeFactory.instance.textNode("Value in English"));
         List<LanguageValue> languageValues = HelperUtil.getLanguageValuesList(arrayNode);
-        assertNotNull(languageValues);
-        assertEquals(1, languageValues.size());
-        assertEquals("en", languageValues.get(0).getLanguage());
-        assertEquals("Value in English", languageValues.get(0).getValue());
+        Assertions.assertNotNull(languageValues);
+        Assertions.assertEquals(1, languageValues.size());
+        Assertions.assertEquals("en", languageValues.get(0).getLanguage());
+        Assertions.assertEquals("Value in English", languageValues.get(0).getValue());
     }
 
     @Test
     public void generateB64EncodedHash_validDetails_thenPass() throws NoSuchAlgorithmException, MockIdentityException {
         String hash = HelperUtil.generateB64EncodedHash("SHA-256", TEST_STRING);
-        assertNotNull(hash);
+        Assertions.assertNotNull(hash);
         MessageDigest digest = MessageDigest.getInstance("SHA-256");
         byte[] expectedHash = digest.digest(TEST_STRING.getBytes(StandardCharsets.UTF_8));
         String expectedEncoded = HelperUtil.b64Encode(expectedHash);
-        assertEquals(expectedEncoded, hash);
+        Assertions.assertEquals(expectedEncoded, hash);
     }
 
     @Test
     public void generateB64EncodedHash_withInvalidAlgorithm_thenFail() {
-        MockIdentityException exception = assertThrows(MockIdentityException.class, () -> {
+        MockIdentityException exception = Assertions.assertThrows(MockIdentityException.class, () -> {
             HelperUtil.generateB64EncodedHash("INVALID_ALGO", TEST_STRING);
         });
-        assertEquals("invalid_algorithm", exception.getMessage());
+        Assertions.assertEquals("invalid_algorithm", exception.getMessage());
     }
 }
